+++ golem/transactions/transactionsystem.py
@@ -1,5 +1,3 @@
-from ethereum.utils import denoms
-
 from golem.model import Payment
 
 from paymentskeeper import PaymentsKeeper
@@ -69,15 +67,3 @@ class TransactionSystem(object):
         # return after_deadline
 
         return []
-
-    def get_payment_for_subtasks(self, subtasks):
-        """
-        Get cost of subtasks
-        :param subtasks: List of subtasks ID
-        :return: Cost of subtasks
-        """
-        cost = 0.0
-        if self.payments_keeper:
-            for subtask in subtasks:
-                cost += self.payments_keeper.get_payment(subtask)
-        return cost / denoms.ether
